+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/react
 
+## 0.0.35
+
+### Patch Changes
+
+- added textarea
+
 ## 0.0.34
 
 ### Patch Changes